requirements.txt
uvicorn
dotenv
fastmcp
langchain_huggingface
httpx
fastapi
python-docx